+++ 26.2/common/src/main/java/com/iamkaf/bonded/leveling/levelers/UtilityToolLeveler.java
@@ -1,0 +1,21 @@
+package com.iamkaf.bonded.leveling.levelers;
+
+import com.iamkaf.bonded.registry.Tags;
+import com.iamkaf.bonded.registry.TierMap;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.*;
+import net.minecraft.world.item.crafting.Ingredient;
+
+import java.util.Objects;
+
+public class UtilityToolLeveler implements GearTypeLeveler {
+    @Override
+    public String name() {
+        return "Utility Tools";
+    }
+
+    @Override
+    public TagKey<Item> tag() {
+        return Tags.UTILITY_EQUIPMENT;
+    }
+}