+++ 26.2/fabric/src/main/java/com/iamkaf/bonded/BondedFabric.java
@@ -1,0 +1,14 @@
+package com.iamkaf.bonded;
+
+import net.fabricmc.api.ModInitializer;
+
+/**
+ * Fabric entry point.
+ */
+public class BondedFabric implements ModInitializer {
+
+    @Override
+    public void onInitialize() {
+        Bonded.init();
+    }
+}