+++ 26.2/fabric/src/main/java/com/iamkaf/bonded/fabric/datagen/ModBlockLootTableProvider.java
@@ -1,0 +1,19 @@
+package com.iamkaf.bonded.fabric.datagen;
+
+import com.iamkaf.bonded.registry.Blocks;
+import java.util.concurrent.CompletableFuture;
+import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootSubProvider;
+import net.minecraft.core.HolderLookup;
+
+public class ModBlockLootTableProvider extends FabricBlockLootSubProvider {
+    public ModBlockLootTableProvider(FabricPackOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+        super(output, registriesFuture);
+    }
+
+    @Override
+    public void generate() {
+        dropSelf(Blocks.TOOL_BENCH.get());
+        dropSelf(Blocks.REPAIR_BENCH.get());
+    }
+}